backend/requirements.txt
# Django 和 REST framework
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.2
django-storages==1.14.2
django-cleanup==8.0.0
django-admin-interface==0.28.0
openpyxl>=3.1.2

# 数据库和缓存
PyMySQL>=1.0.3
redis>=5.0.0
django-redis==5.4.0

# 图片处理
Pillow==10.0.0

# 工具包
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==23.3.0
whitenoise==6.5.0

# 微信支付
wechatpy==1.8.18
pycryptodome>=3.19.0
xmltodict>=0.13.0

# 开发工具
django-debug-toolbar==4.1.0

# 其他依赖
mysqlclient==2.2.0
boto3==1.34.69
django-extensions==3.2.3
django-silk==4.3.0
django-import-export==3.3.1
django-simple-history==3.4.0
django-widget-tweaks==1.5.0
django-ckeditor==6.7.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
celery==5.3.4
qrcode==7.4.2
cryptography==41.0.3
requests==2.31.0
python-dateutil==2.8.2 